Get posts for current Page
Created function that gets all posts, filters them down to only the
posts that have a parent page id that is the same as the current page.
It works!
Now we are adding event handlers and functionality to the
"create new post" button. This will bring up create post fields.

This was an excellent group code.