Builds visitor view for non-auth users to see content
Creates handlebars file to parse and load get request for all pages.
Creates functions in events/ui/index/layouts/api/events/ui in order
to correctly display the visitor content, as well as set click handlers
so that when a user logs in they can view pages and posts correctly.