Removes cancel buttons, fixes error for posts
Removes cancel buttons from around the app. Were causing bugs and breaks
in functionality and aren't needed so they are commented out in this
commit.

Backend repo now has a fix for posting to posts that aren't yours, so
the error here is updated to be consistent with user experience.