Updates event for getting pages, and changes some messages
API function was expecting _owner to be passed, but it wasn't. This
adds it from store when the API is fired from the event.

Also changes/removes some messages that were causing bad UI.

***NOTE*****
This is getting a "not allowed by Access-Control-Allow-Methods" now
on my local.

Going to debug but not going to merge to master.
***NOTE****